requirements.txt
# Core
fastapi>=0.110.0
uvicorn>=0.29.0
apscheduler>=3.10.4
python-dotenv>=1.0.1

# YouTube Music
ytmusicapi>=1.11.0

# Spotify
spotipy>=2.25.0

# Matching
thefuzz>=0.22.1
python-Levenshtein>=0.25.1

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
